requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
numpy==2.2.4
packaging==24.2
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pyqtgraph==0.13.7
pyserial==3.5
